change directories position
